leaffn.py: a third reader for the straight-line leaves, 845/2172 (38.90%)
A census of what is left in gamecode/logic found a hundred functions under 96
bytes with no branch in them -- no guard, no handler, just stores into the
object and calls. pushfn refuses every one for want of a guard it was built to
require, so tools/leaffn.py is that same interpreter with the requirement
dropped and the OBJECT in r0.

22 written. tools/leafall.py sweeps every file in one process, because
genfile.py rebuilds the symbol table per reader per file and that is far too
slow for a sweep.

**Four refusals, each added because guessing there produced something wrong.**
All four were caught by reading one output against dumpfn.py, not by reasoning
about the tool:

  A value read from a field the function also writes is a SAVED value being
  put back, not a re-read. setup_proj_obj loads 0x40, overwrites it, calls
  out, and restores the original -- written as a re-read that becomes
  `obj->field40 = obj->field40`, which it did.

  A function is `void` only when the last thing to touch r0 was PROVABLY a
  call, so what is left there is the callee's leftover. This is the same
  mistake that hid randper's return and DrawSkinnedMesh2's count.

  A call is refused when r2 or r3 were written for it, because that is a third
  or fourth argument this reader does not model. secret_move_search takes
  three and came out with one.

  A call whose r1 was loaded for it takes a second argument, and it is
  rendered or the function is refused. q_am_i_a_boss is `ldr r1, [r0, #8]`
  then `bl bossck` -- bossck(obj, obj->field08) -- and came out as bossck(obj).

Two smaller fixes on the way: genfile.py had a real newline inside a string
literal from an earlier heredoc, so it had not run since; and a definition
arriving now CORRECTS an earlier reader's guessed return type in place rather
than being dropped, because the callers above it still need the declaration.

Also closes two of the nine in #26: HUDANIM and EPLAYER are both `typedef int`,
so the declarations saying `long` just needed to say `int`.

docs/HANDOFF.md now describes all three readers and, more usefully, what each
refuses and why.

tree: 0 errors. Spot-checked run_setup, d_behind_me_a5 and
last_knockdown_frame against the disassembly: exact.

+++ decomp/gamecode/Blood.c
@@ -616,7 +616,9 @@ void  SaveUnclaimedTreasure(long treasure);
 void  DumpAltCostume(char *player);
 void  LoadGameCharacterCheckCache(char *player, const PLAYERDEF *def,
                                   long *stats);
-void  HUDANIM_TriggerAnim(long anim);
+/* HUDANIM is `typedef int` in HudAnim.c, so int is that same type rather
+ * than a second copy of the typedef here. */
+void  HUDANIM_TriggerAnim(int anim);
 void  StopCameraTracking(void);
 void  SetCameraOverridePosFrom2d(float x, float y, float z);
 void  limeMatrixLoadIdentity(float *mtx);
